Add menu navigation on back press in fragments
A new functionality has been added in CityFragment and PersonFragment to handle back press. Now when back button is pressed, the application will navigate the user back to the MenuFragment. This enhancement improves user navigation experience by providing a clear path to return to the main menu from any fragment. The changes necessitated importing the MenuFragment class in both CityFragment and PersonFragment files.